% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.odrl_fit}
\alias{fitted.odrl_fit}
\title{Extract fitted treatment recommendations}
\usage{
\method{fitted}{odrl_fit}(object, ...)
}
\arguments{
\item{object}{An \code{\link[=odrl]{odrl()}} fit.}

\item{...}{Unused.}
}
\value{
Treatment recommendations in the original treatment coding.
}
\description{
Extract fitted treatment recommendations
}
