% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.odrl_fit}
\alias{print.odrl_fit}
\title{Print an ODRL fit}
\usage{
\method{print}{odrl_fit}(x, ...)
}
\arguments{
\item{x}{An \code{\link[=odrl]{odrl()}} fit.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print an ODRL fit
}
