% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.odrl_fit}
\alias{summary.odrl_fit}
\title{Summarize an ODRL fit}
\usage{
\method{summary}{odrl_fit}(object, ...)
}
\arguments{
\item{object}{An \code{\link[=odrl]{odrl()}} fit.}

\item{...}{Unused.}
}
\value{
An object of class \code{summary.odrl_fit}.
}
\description{
Summarize an ODRL fit
}
