% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odrlITR-package.R
\docType{package}
\name{odrlITR-package}
\alias{odrlITR}
\alias{odrlITR-package}
\title{Orthogonal Double Residual Learning}
\description{
The \code{odrlITR} package learns binary individualized treatment rules from the
cross-fitted signed score
\deqn{\widehat Z_i = \{A_i-\widehat e(X_i)\}
      \{Y_i-\widehat m(X_i)\},}
where treatment is internally coded as \eqn{A\in\{-1,+1\}},
\eqn{e(x)=E(A\mid X=x)}, and \eqn{m(x)=E(Y\mid X=x)}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/deckardt98/odrlITR}
  \item Report bugs at \url{https://github.com/deckardt98/odrlITR/issues}
}

}
\author{
\strong{Maintainer}: Jiaqi Tong \email{jiaqi.tong@yale.edu} (\href{https://orcid.org/0009-0005-8922-3386}{ORCID})

Authors:
\itemize{
  \item Fan Li
}

}
\keyword{internal}
